savings system tweak

+++ include/script_menu.h
@@ -15,9 +15,9 @@
 extern const u8 *const gStdStrings[];
 
 struct NumericInput {
-    u16 min;
-    u16 max;
-    u16 value;
+    u32 min;
+    u32 max;
+    u32 value;
     u8 selectedDigit;
 };
 
@@ -56,7 +56,7 @@ int ScriptMenu_AdjustLeftCoordFromWidth(int left, int width);
 bool16 ScriptMenu_CreatePCMultichoice(void);
 void ScriptMenu_DisplayPCStartupPrompt(void);
 u32 CreateNumericInputWindow(s16 x, s16 y, u8 width, u8 numDigits, const u8* templStr);
-void SetNumericInputDefault(u16 min, u16 max, u16 initial);
+void SetNumericInputDefault(u32 min, u32 max, u32 initial);
 void PrintNumericInputAmount(u8 windowId, const u8* templString);
 bool32 HandleNumericInput(void);
